requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.3.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.6
Django==4.0
django-celery-results==2.5.1
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-phone-field==1.8.1
executing==1.2.0
export==0.2.1
Faker==18.11.2
filelock==3.12.2
flake8==6.0.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
kombu==5.3.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
pluggy==1.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
pygraphviz==1.11
pyproject-api==1.5.2
python-dateutil==2.8.2
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tox==4.6.3
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
virtualenv==20.23.1
wcwidth==0.2.6
